Move integration and e2e tests for epp into epp-specific directories
