+++ test/e2e/epp/e2e_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e
+package epp
 
 import (
 	"fmt"
